added normaladapter and implementation in ow
there are now 2 adapter, a dummy adapter, which is for testing purposes
and a normal adapter, which is used for open wonderland.

+++ src/org/jdesktop/wonderland/modules/oweditor/client/adapter/ClientUpdateAdapter.java
@@ -9,7 +9,7 @@ public class ClientUpdateAdapter implements ClientUpdateInterface{
     }
 
     @Override
-    public void updateTranslation(int id, int x, int y, int z) {
+    public void updateTranslation(long id, int x, int y, int z) {
         mc.getSUA().serverChangeEvent(id, x, y, 0, 0, 1, "");
         
     }
