+++ src/org/jdesktop/wonderland/modules/oweditor/client/adapter/ClientUpdateInterface.java
@@ -2,6 +2,6 @@ package org.jdesktop.wonderland.modules.oweditor.client.adapter;
 
 public interface ClientUpdateInterface {
     
-    public void updateTranslation(int id, int x, int y, int z);
+    public void updateTranslation(long id, int x, int y, int z);
 
 }
