+++ src/org/jdesktop/wonderland/modules/oweditor/client/editor/gui/ShapeObjectSelectionRect.java
@@ -42,7 +42,7 @@ public class ShapeObjectSelectionRect extends ShapeObject{
     }
 
     @Override
-    public int getID() {
+    public long getID() {
         return 0;
     }
